@@ -5,7 +5,7 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"log"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"os"
@@ -30,7 +30,7 @@ func (s *BotTestSuite) SetupSuite() {
 		Options: Options{
 			limit:      100,
 			timeout:    25,
-			logger:     newLogger("micha"),
+			logger:     slog.Default(),
 			apiServer:  defaultAPIServer,
 			httpClient: http.DefaultClient,
 		},
@@ -132,10 +132,10 @@ func (s *BotTestSuite) TestNewBot() {
 	s.Require().NotNil(bot)
 	s.Require().Equal(25, bot.timeout)
 	s.Require().Equal(100, bot.limit)
-	s.Require().Equal(newLogger("[micha] "), bot.logger)
+	s.Require().Equal(slog.Default(), bot.logger)
 
 	// With options
-	logger := log.New(os.Stderr, "", log.LstdFlags)
+	logger := slog.New(slog.NewTextHandler(os.Stderr, nil))
 	httpClient := &http.Client{}
 	bot, err = NewBot("111", WithLimit(50), WithTimeout(10), WithLogger(logger), WithHttpClient(httpClient))
 	s.Require().Nil(err)
